requirements.txt
# FastAPI and web
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart
PyPDF2>=0.0.18
aiofiles>=24.1.0

# Database
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
alembic>=1.14.0

# LangChain and AI
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0

# Vector database
chromadb>=0.5.0
langchain-chroma>=0.1.2
langchain-text-splitters>=0.3.0

# Document processing
pypdf>=5.1.0
python-docx>=1.1.2
pytesseract>=0.3.13
pillow>=11.0.0
pdf2image>=1.17.0

# Voice (modern Python 3.13 compatible)
faster-whisper>=1.1.0
soundfile>=0.12.1

# Embeddings
sentence-transformers>=3.3.1

# Validation and types
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Dependency injection
dependency-injector>=4.43.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
pytest-mock>=3.14.0
httpx>=0.28.0

# Development
ruff>=0.8.0
black>=24.10.0
mypy>=1.13.0

# Authentication
argon2-cffi>=23.1.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4

# Utilities
python-dotenv>=1.0.1
loguru>=0.7.3
aiohttp>=3.9.0
python-json-logger>=2.0.7
tldextract>=5.3.0
